Replaced int64_t with int in BC_ScrollBar
There will not be screens what pixel count needs 64bit
in a foreseeable future. Some cosmetics.
Use SPDX-License tag.

+++ guicast/bcscrollbar.h
@@ -1,23 +1,7 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
 
-/*
- * CINELERRA
- * Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
- * 
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- * 
- */
+// This file is a part of Cinelerra-CVE
+// Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
 
 #ifndef BCSCROLLBAR_H
 #define BCSCROLLBAR_H
@@ -30,43 +14,31 @@
 #define SCROLL_VERT  1
 
 // Selection identifiers
-#define SCROLL_HANDLE 1
-#define SCROLL_BACKPAGE 2
-#define SCROLL_FWDPAGE 3
-#define SCROLL_BACKARROW 4
-#define SCROLL_FWDARROW 5
+#define SCROLL_HANDLE     1
+#define SCROLL_BACKPAGE   2
+#define SCROLL_FWDPAGE    3
+#define SCROLL_BACKARROW  4
+#define SCROLL_FWDARROW   5
 
 // Image identifiers
-#define SCROLL_HANDLE_UP 	0
-#define SCROLL_HANDLE_HI 	1
-#define SCROLL_HANDLE_DN 	2
-#define SCROLL_HANDLE_BG 	3
-#define SCROLL_BACKARROW_UP 4
-#define SCROLL_BACKARROW_HI 5
-#define SCROLL_BACKARROW_DN 6
-#define SCROLL_FWDARROW_UP  7
-#define SCROLL_FWDARROW_HI  8
-#define SCROLL_FWDARROW_DN  9
-#define SCROLL_IMAGES 		10
-
-
-
-
-
-
-
+#define SCROLL_HANDLE_UP      0
+#define SCROLL_HANDLE_HI      1
+#define SCROLL_HANDLE_DN      2
+#define SCROLL_HANDLE_BG      3
+#define SCROLL_BACKARROW_UP   4
+#define SCROLL_BACKARROW_HI   5
+#define SCROLL_BACKARROW_DN   6
+#define SCROLL_FWDARROW_UP    7
+#define SCROLL_FWDARROW_HI    8
+#define SCROLL_FWDARROW_DN    9
+#define SCROLL_IMAGES         10
 
 
 class BC_ScrollBar : public BC_SubWindow
 {
 public:
-	BC_ScrollBar(int x, 
-		int y, 
-		int orientation, 
-		int pixels, 
-		int64_t length, 
-		int64_t position, 
-		int64_t handlelength,
+	BC_ScrollBar(int x, int y, int orientation, int pixels,
+		int length, int position, int handlelength,
 		VFrame **data = 0);
 	virtual ~BC_ScrollBar();
 
@@ -80,15 +52,15 @@ public:
 	int button_press_event();
 	int button_release_event();
 	void repeat_event(int repeat_id);
-	int64_t get_value();
-	int64_t get_position();
-	int64_t get_length();
-	int64_t get_handlelength();
+	int get_value();
+	int get_position();
+	int get_length();
+	int get_handlelength();
 	int get_pixels();
 	void set_images(VFrame **data);
 	int in_use();
-	void update_value(int64_t value);
-	void update_length(int64_t length, int64_t position, int64_t handlelength);
+	void update_value(int value);
+	void update_length(int length, int position, int handlelength);
 	void reposition_window(int x, int y, int pixels);
 	int get_span();
 	static int get_span(int orientation);
@@ -101,24 +73,16 @@ private:
 	void get_handle_dimensions();
 	int get_cursor_zone(int cursor_x, int cursor_y);
 
-	int64_t length, position, handlelength;   // handle position and size
+	int length, position, handlelength;   // handle position and size
 	int selection_status, highlight_status;
 	int orientation, pixels;
 	int handle_pixel, handle_pixels;
 	int min_pixel, max_pixel;
-	int64_t repeat_count;
+	int repeat_count;
 // Don't deactivate if bound to another tool
 	BC_WindowBase *bound_to;
 	VFrame **data;
 	BC_Pixmap *images[SCROLL_IMAGES];
 };
-
-
-
-
-
-
-
-
 
 #endif
